src/main/resources/python/requirements.txt
# YOLOv8 水果检测依赖
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
